[fix] handle Alpenglow u64 epoch credits
Bump gagliardetto/solana-go v1.8.4 -> v1.20.0 so getVoteAccounts decodes
EpochCredits as uint64, preventing the int64 parse overflow caused by
Alpenglow's u64::MAX sentinel values in delinquent validators' credits.

Credit-ranking and credit-sample code now compute the per-epoch diff in
uint64 and guard against the math.MaxUint64 sentinel (and underflow),
treating those as zero credits so affected accounts sink to the bottom
of the ranking instead of overflowing.

The uint64 EpochCredits fix is not yet released upstream
(solana-foundation/solana-go#486), so this temporarily points at a fork
via a replace directive pinned to that PR's commit:

    replace github.com/gagliardetto/solana-go => github.com/mindrunner/solana-go v0.0.0-20260830083219-f93c6c7c1a64

The replace will be dropped and the requirement repinned once the fix
ships in a tagged solana-go release.

+++ internal/solana/mock_client.go
@@ -251,8 +251,8 @@ func (b *MockClientBuilder) WithVoteAccount(pubkey string, rank int, credits int
 		if p == pubkey {
 			return &rpc.VoteAccountsResult{
 				NodePubkey: solana.MustPublicKeyFromBase58(pubkey),
-				EpochCredits: [][]int64{
-					{1, credits, credits / 2},
+				EpochCredits: [][]uint64{
+					{1, uint64(credits), uint64(credits) / 2},
 				},
 			}, rank, nil
 		}
